requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Configuration
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Date/Time
pytz==2025.2

# Core Python Dependencies
asgiref==3.8.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2

# Data Processing (if needed for analytics)
numpy==2.2.6
pandas==2.3.0

# Optional: Image Processing (if used)
Pillow==10.1.0

# Optional: Health Checks
requests==2.31.0